+++ src/utils/theme.js
@@ -152,9 +152,7 @@ export default function setTheme(theme = {}) {
     --code-bg:${newTheme.codeBg};
     --code-fg:${newTheme.codeFg};
     --inline-code-fg:${newTheme.inlineCodeFg};
-    --code-property-color:${newTheme.codePropertyColor};
-    --code-keyword-color:${newTheme.codeKeywordColor};
-    --code-operator-color:${newTheme.codeOperatorColor};
+    
 
     /* Computed Color properties */
     --primary-color: ${theme.primaryColor};