+++ CHANGELOG.md
@@ -17,6 +17,7 @@ This package follows standard semvar, `<major>.<minor>.<build>`. No breaking cha
 * Add support for `query` REST verb type
 * Hide empty oneof empty clauses that don't affect displayed properties.
 * Fix schema table transitions and display
+* Fix response code colors to match the rest of the theme
 
 ## 0.9 ##
 * OpenAPI 3.1 support using `APIDevTools/json-schema-ref-parser`